Split download and assembly components
