+++ debootstrap.py
@@ -3,7 +3,6 @@
 * Right now LZMA decoding takes up most of the time. Parallelize it? Python's LZMA
   library does release the GIL.
 """
-import fnmatch
 import json
 import gzip
 import lzma
@@ -36,90 +35,13 @@ from tempfile import NamedTemporaryFile
 from urllib.parse import urlparse
 from wsgiref.handlers import format_date_time
 
-try:
-    from zstandard import ZstdDecompressor
-except ModuleNotFoundError:
-    pass
+from lib.assemble import create_filesystem
 
 NUL = b"\0"
 BLOCKSIZE = tarfile.BLOCKSIZE
 CACHE_PATH = Path("debs")
 GNUPG_PREFIX = b"[GNUPG:] "
 PACKAGES_PREFERENCE = {".xz": lzma.open, ".gz": gzip.open, "": lambda f, mode: f}
-
-
-class GPGVNotFoundError(Exception):
-    pass
-
-
-def stderr(*args, **kwargs):
-    kwargs["file"] = sys.stderr
-    return print(*args, **kwargs)
-
-
-@classmethod
-def zstdopen(cls, name, mode="r", fileobj=None, **kwargs):
-    dctx = ZstdDecompressor()
-    fileobj = dctx.stream_reader(fileobj)
-    try:
-        t = cls.taropen(name, mode, fileobj, **kwargs)
-    except:  # noqa: E722
-        fileobj.close()
-        raise
-
-    t._extfileobj = False
-    return t
-
-
-tarfile.TarFile.zstdopen = zstdopen
-tarfile.TarFile.OPEN_METH["zstd"] = "zstdopen"
-
-
-SECOND_STAGE = r"""#!/bin/bash
-set -e
-
-cat << EOF > /usr/bin/policy-rc.d
-#!/bin/sh
-exit 101
-EOF
-chmod 755 /usr/bin/policy-rc.d
-
-echo "Making control file" >&2
-cd /var/lib/dpkg/info
-for f in *.control; do
-  cat $f
-  echo
-done > /var/lib/dpkg/status
-rm -r *.control
-
-# SOURCE_DATE_EPOCH makes /etc/shadow reproducible
-export DEBIAN_FRONTEND=noninteractive SOURCE_DATE_EPOCH=0
-
-set -x
-for script in *.preinst; do
-  package_fullname="${script//.preinst}"
-  package_name="${package_fullname//:*}"
-  DPKG_MAINTSCRIPT_NAME=preinst \
-  DPKG_MAINTSCRIPT_PACKAGE=$package_name \
-  ./"$script" install
-done
-
-cd /
-# libc6's postinst requires `which`, which is configured via update-alternatives(1)
-dpkg --configure --force-depends debianutils
-dpkg --configure -a
-
-rm /etc/passwd- /etc/group- /etc/shadow- \
-  /var/cache/debconf/*-old /var/lib/dpkg/*-old \
-  /init
-# This cache is not reproducible
-rm /var/cache/ldconfig/aux-cache
-# Some log files (e.g. btmp) need to exist with the right modes, so we truncate them
-# instead of deleting them.
-find /var/log -type f -exec truncate -s0 {} \;
-"""
-ADD_SOURCES_LIST = "echo deb {archive_url} {suite} main >> /etc/apt/sources.list\n"
-
 THIRD_STAGE = r"""
 # Make suitable for VM use
 passwd -d root
@@ -140,6 +62,16 @@ EOF
 """
 
 
+
+class GPGVNotFoundError(Exception):
+    pass
+
+
+def stderr(*args, **kwargs):
+    kwargs["file"] = sys.stderr
+    return print(*args, **kwargs)
+
+
 FIELDS = (
     "Package",
     "Filename",
@@ -150,17 +82,7 @@ FIELDS = (
     "Pre-Depends",
 )
 FIELDS_MATCHER = re.compile("^({}): (.*)$".format("|".join(FIELDS)))
-LOCALE_MATCHER = re.compile(fnmatch.translate("usr/share/locale/*/LC_MESSAGES/*.mo"))
 
-
-def is_excluded(name):
-    if name.startswith("usr/share/doc/"):
-        return True
-
-    if name.startswith("usr/share/man/"):
-        return True
-
-    return bool(LOCALE_MATCHER.match(name))
 
 
 def packages_dict(packages):
@@ -244,9 +166,7 @@ def copy_file_sha256(src, dst):
     dst.flush()
     return hasher.hexdigest()
 
-
 threadlocals = threading.local()
-
 
 def download_file(netloc, url, out_fh):
     r = fetch_http(netloc, url)
@@ -254,208 +174,6 @@ def download_file(netloc, url, out_fh):
         raise RuntimeError(r.status)
 
     return copy_file_sha256(r, out_fh)
-
-
-WANTED_LINES = set(["Package", "Architecture", "Multi-Arch"])
-
-
-def _get_dpkg_name(control):
-    if control.get("Multi-Arch", None) == "same":
-        return "{}:{}".format(control["Package"], control["Architecture"])
-
-    return control["Package"]
-
-
-def parse_control_data(data):
-    lines = data.splitlines(True)
-    parsed = dict()
-    for idx, line in enumerate(lines):
-        parts = line.split(": ", 1)
-        if len(parts) != 2:
-            continue
-
-        k, v = parts
-        if k in WANTED_LINES:
-            parsed[k] = v.rstrip()
-
-    # the table at lib/dpkg/parse.c seems to determine the "correct" order
-    # of fields, but we just drop this last.
-    # We assume that the next run of dpkg will fix it (it does)
-    lines.append("Status: install ok unpacked\n")
-
-    name = _get_dpkg_name(parsed)
-    return ("var/lib/dpkg/info/{}.".format(name), "".join(lines).encode())
-
-
-def _dpkg_info_files(prefix, control_data, tf):
-    control_info = TarInfo(prefix + "control")
-    control_info.size = len(control_data)
-    yield control_info, control_data
-
-    for member, file_contents in tf:
-        if not member.isreg():
-            continue
-
-        name = member.name.lstrip("./")
-        if name == "control":
-            continue
-        member.name = prefix + name
-        yield member, file_contents
-
-
-def extract_whole_tar(contents):
-    tf = tarfile.open(fileobj=BytesIO(contents))
-    ret = dict()
-    for ti in tf:
-        inner_fh = tf.extractfile(ti)
-        file_data = None if inner_fh is None else inner_fh.read()
-        ret[ti.name] = (ti, file_data)
-    return ret
-
-
-def handle_control_tar(contents):
-    data = extract_whole_tar(contents)
-    control_data = data["./control"][1].decode()
-    prefix, new_control_data = parse_control_data(control_data)
-    return prefix, _dpkg_info_files(prefix, new_control_data, data.values())
-
-
-def transform_name(name):
-    if name == "":
-        # I have no idea why dpkg does this
-        return "/.\n"
-
-    return "/" + name + "\n"
-
-
-def unpack_ar(fh):
-    assert fh.read(8) == b"!<arch>\n"
-    prefix = None
-    files = []
-    while True:
-        header = fh.read(60)
-        if not header:
-            break
-
-        name = header[0:16]
-
-        size = int(header[48:58], 10)
-        assert header[58:60] == b"\x60\x0A"
-        file_contents = fh.read(size)
-        fh.read(size % 2)
-
-        if name.startswith(b"data.tar"):
-            tf = tarfile.open(fileobj=BytesIO(file_contents))
-            for member in tf:
-                contents = tf.extractfile(member).read() if member.isreg() else None
-                member.name = member.name.lstrip("./")
-                files.append(transform_name(member.name))
-                yield member, contents
-
-        if name.startswith(b"control.tar"):
-            prefix, dpkg_files = handle_control_tar(file_contents)
-            yield from dpkg_files
-
-    if prefix is None:
-        raise RuntimeError("Missing control file?")
-
-    # This becomes the dpkg .list info file
-    files_manifest = "".join(files).encode()
-    info = TarInfo(prefix + "list")
-    info.size = len(files_manifest)
-    yield info, files_manifest
-
-
-class Filesystem:
-    def __init__(self):
-        self._files = dict()
-
-    def mkdir(self, name):
-        ti = TarInfo(name)
-        ti.type = tarfile.DIRTYPE
-        ti.mode = 0o755
-        self.add(ti)
-
-    def symlink(self, name, target):
-        ti = TarInfo(name)
-        ti.type = tarfile.SYMTYPE
-        ti.linkname = target
-        self.add(ti)
-
-    def file(self, name, contents, mode=None):
-        ti = TarInfo(name)
-
-        if isinstance(contents, str):
-            contents = contents.encode()
-
-        ti.size = len(contents)
-        if mode is not None:
-            ti.mode = mode
-        self.add(ti, BytesIO(contents))
-
-    def mknod(self, name, major, minor):
-        ti = TarInfo(name)
-        ti.type = tarfile.CHRTYPE
-        ti.devmajor = major
-        ti.devminor = minor
-        self.add(ti)
-
-    def resolve(self, name):
-        try:
-            entry = self._files[name]
-        except KeyError:
-            return name
-
-        info, fh = entry
-        if not info.issym():
-            return name
-
-        dirname = path.dirname(name)
-        target = path.normpath(path.join(dirname, info.linkname))
-        return self.resolve(target)
-
-    def _build_path(self, name):
-        ret = ""
-        components = name.split("/")[::-1]
-        while components:
-            c = components.pop()
-            ret = self.resolve(path.join(ret, c))
-
-        return ret
-
-    def add(self, ti, fileobj=None):
-        ti.name = self._build_path(ti.name)
-        ti.uname = ""
-        ti.gname = ""
-
-        if ti.name in self._files:
-            existing, _ = self._files[ti.name]
-            existing.mtime = max(existing.mtime, ti.mtime)
-
-            if extract_useful(ti) != extract_useful(existing):
-                raise RuntimeError(ti.name)
-
-            return
-
-        if ti.name == "":
-            return
-
-        self._files[ti.name] = (ti, fileobj)
-
-
-def extract_useful(ti):
-    return (
-        ti.name,
-        ti.mode,
-        ti.uid,
-        ti.gid,
-        ti.size,
-        ti.type,
-        ti.uname,
-        ti.gname,
-        ti.pax_headers,
-    )
-
 
 def download_files(parsed_archive_url, packages):
     executor = ThreadPoolExecutor(8)
@@ -484,44 +202,6 @@ def download_files(parsed_archive_url, packages):
         os.link(temp_fh.name, destination)
         stderr(f"Downloaded {destination}")
         yield destination
-
-
-def get_debs_from_directory(paths):
-    for deb in paths:
-        yield deb.open("rb")
-
-
-def get_unpacked_files(fhs):
-    for fh in fhs:
-        yield from unpack_ar(fh)
-        fh.close()
-
-
-def create_filesystem(deb_names, add_sources_list: str):
-    fs = Filesystem()
-
-    second_stage = SECOND_STAGE
-    for info in add_sources_list:
-        second_stage += ADD_SOURCES_LIST.format(**info)
-
-    second_stage += THIRD_STAGE
-
-    fs.file("init", second_stage, mode=0o755)
-
-    # These files will get created by dpkg as well..mostly.
-    # There's a risk that zero packages will install these files.
-    # Which is why we create them manually
-    for name in ("bin", "sbin", "lib", "lib32", "lib64", "libx32"):
-        real = f"usr/{name}"
-        fs.mkdir(real)
-        fs.symlink(name, real)
-
-    debs = get_debs_from_directory(deb_names)
-    for member, contents in get_unpacked_files(debs):
-        fs.add(member, BytesIO(contents))
-
-    return fs
-
 
 def pretty_time(value):
     if value < 0.001:
@@ -602,24 +282,6 @@ class SHA256File:
         return self._write_timer.value
 
 
-def write_file(out_fh, info, fh):
-    if not info.isdir() and is_excluded(info.name):
-        return
-
-    out_fh.write(info.tobuf())
-    tarfile.copyfileobj(fh, out_fh, info.size)
-    blocks, remainder = divmod(info.size, BLOCKSIZE)
-    if remainder == 0:
-        return
-
-    out_fh.write(NUL * (BLOCKSIZE - remainder))
-
-
-def write_image(fs, out_fh):
-    files = fs._files
-    for name in sorted(files):
-        write_file(out_fh, *files[name])
-
 
 class NullFile:
     @staticmethod
@@ -644,8 +306,7 @@ def mutate_file(fs, ti):
     original_entry = fs._files.get(ti.name)
     if original_entry:
         original_mtime = original_entry[0].mtime
-        if original_mtime != ti.mtime:
-            ti.mtime = original_mtime
+        ti.mtime = original_mtime
     else:
         ti.mtime = 0
 
@@ -931,14 +592,14 @@ def build_os(*, architecture, keyring, archive_url, suites):
     stderr("Creating filesystem")
     deb_paths = download_files(parsed_archive_url, packages)
     sources_entries = [dict(archive_url=archive_url, suite=suite) for suite in suites]
-    fs = create_filesystem(deb_paths, add_sources_list=sources_entries)
+    fs = create_filesystem(deb_paths, add_sources_list=sources_entries, third_stage=THIRD_STAGE)
 
     stderr("Writing image to docker import")
     docker_import_p = Popen(["docker", "import", "-"], stdin=PIPE, stdout=PIPE)
     with docker_import_p.stdin as fh:
         hasher = SHA256File(fh)
         with Timer() as timer:
-            write_image(fs, hasher)
+            fs.write(hasher)
     timer.value -= hasher.hash_time + hasher.write_time
     stderr(f"Hashing took {pretty_time(hasher.hash_time)} seconds")
     stderr(f"Writing took {pretty_time(hasher.write_time)} seconds")
